oculomotor/requirements.txt
numpy==1.14.5
flask
flask-cors
pygame
pyglet
opencv-python
opencv-contrib-python
pybind11
BriCA2
oculoenv==0.1.3
tensorflow
scipy
chainer
torch
torchvision
